requirements.txt
colorama==0.4.6
crccheck==1.3.0
Pillow==9.5.0
pyasn1==0.5.0
pypng==0.20220715.0
pyzbar==0.1.9
qrcode==7.4.2
rsa==4.9
typing_extensions==4.5.0
